backend/requirements.txt
django
djangorestframework
psycopg2-binary
firebase-admin
boto3
django-cors-headers
botocore